chore: one less file
